requirements.txt
flet~=0.28.3
python-dotenv~=1.1.1
google-genai
protobuf~=5.29.5
supabase~=2.16.0
customtkinter~=5.2.2
pillow~=11.3.0